Ensure all roff-fonts are usable by ghostscript
Fonts included in gyre-fonts have filenames following pattern
"texgyrefont-style.otf" however the "actual" name of the font is
"TeXGyreFont-Style". The latter name is what ends up in the postscript
generated by roff | post; ghostscript is then unable to find the font
files when processing the postscript.

This can be addressed by creating a Fontmap file of the format:

  /HumanFontName1 (font_file_path1) ;
  /HumanFontName2 (font_file_path2) ;
  ...
  /HumanFontNameN (font_file_pathN) ;

This is now generated during the build of roff-fonts package and placed
at /share/neatroff/Fontmap. Additionally, ghostscript is now included in
the main neatroff package, and the "gs" binary is wrapped to add flag
"-I${roff-fonts}/share/neatroff so it can find the Fontmap file.

One minor change also included is removing the use of neatmkfn's gen.sh
script inside roff-fonts build. It was only used to generate the various
"standard" troff font descriptions for ghostscript fonts. Instead the
same logic is now directly in the Makefile for better parallelism and
faster builds.